+++ artifact/JettonTact_JettonDefaultWallet.code.boc
