+++ src_ts/utils/static.ts
